Implement a stash for keys
This stash stores the key in a static fixed-size array that is used like a ring buffer. This is a fast path
for getting the keys, as if the keys are already generated, you will not have to wait around 7-15 seconds generating them. It is run in the
background initially and then only regenerates at index where the key was taken from.